+++ projects/np-ui-lib/src/lib/np-sidepanel/np-sidepanel-content.directive.ts
@@ -1,0 +1,7 @@
+import { Directive } from '@angular/core';
+
+@Directive({
+    selector: '[npSidepanelContent]',
+})
+export class NpSidepanelContent {
+}