+++ projects/np-ui-lib/src/lib/np-sidepanel/np-sidepanel.module.ts
@@ -1,10 +1,12 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { NpSidepanelComponent } from './np-sidepanel.component';
+import { PortalModule } from '@angular/cdk/portal';
+import { NpSidepanelContent } from './np-sidepanel-content.directive';
 
 @NgModule({
-  declarations: [NpSidepanelComponent],
-  imports: [CommonModule],
-  exports: [NpSidepanelComponent]
+  declarations: [NpSidepanelComponent, NpSidepanelContent],
+  imports: [CommonModule, PortalModule],
+  exports: [NpSidepanelComponent, NpSidepanelContent]
 })
 export class NpSidepanelModule { }